requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
django-cors-headers==4.3.1
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
drf-spectacular==0.27.2
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
reportlab==4.2.2
django-allauth==0.61.1

cryptography>=42.0.0
